CIV-12879 Move getClaimTimeline to generic service
